+++ src/components/WeaponsCardWeaponStats.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import WeaponDataRow from './WeaponDataRow';
 import ButtonDeleteX from '../helpers/buttons/ButtonDeleteX';
-import WeaponsCardModifyWeapon from './WeaponsCardModifyFirearm';
+import WeaponsCardModifyWeapon from './WeaponsCardModifyWeapon';
 import { buildArrayForGunTable } from '../helpers/componentHelpers';
 import './WeaponsCard/WeaponsCard.css';
